# web/swagger/dist/doc.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a FFmpeg Go server.",
    "title": "FFmpeg Go API",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "version": "1.0"
  },
  "host": "localhost:8084",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/video/edit": {
      "post": {
        "description": "提交一个新的视频编辑任务",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "video"
        ],
        "summary": "提交视频编辑任务",
        "operationId": "submit-video-edit",
        "parameters": [
          {
            "description": "视频编辑请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VideoEditRequest"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "任务提交成功",
            "schema": {
              "$ref": "#/definitions/VideoEditResponse"
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/video/edit/{id}": {
      "delete": {
        "description": "根据任务ID取消视频编辑任务",
        "produces": [
          "application/json"
        ],
        "tags": [
          "video"
        ],
        "summary": "取消视频编辑任务",
        "operationId": "cancel-video-edit",
        "parameters": [
          {
            "type": "string",
            "description": "任务ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "任务取消成功",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "任务未找到",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "get": {
        "description": "根据任务ID获取视频编辑任务的状态信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "video"
        ],
        "summary": "获取视频编辑任务状态",
        "operationId": "get-video-edit-status",
        "parameters": [
          {
            "type": "string",
            "description": "任务ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "任务状态信息",
            "schema": {
              "$ref": "#/definitions/TaskStatusResponse"
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "任务未找到",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/video/url": {
      "post": {
        "description": "通过URL下载视频并提交处理任务",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "video"
        ],
        "summary": "处理视频URL",
        "operationId": "handle-video-url",
        "parameters": [
          {
            "description": "视频URL请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VideoURLRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "处理成功",
            "schema": {
              "$ref": "#/definitions/VideoURLResponse"
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/workerpool/resize": {
      "post": {
        "description": "动态调整工作池中工作线程的数量",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "workerpool"
        ],
        "summary": "调整工作池大小",
        "operationId": "resize-worker-pool",
        "parameters": [
          {
            "description": "工作池大小调整请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "size": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "工作池调整成功",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/workerpool/status": {
      "get": {
        "description": "获取当前工作池的状态信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "workerpool"
        ],
        "summary": "获取工作池状态",
        "operationId": "get-worker-pool-status",
        "responses": {
          "200": {
            "description": "工作池状态信息",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "OSSOutput": {
      "type": "object",
      "properties": {
        "accessKey": {
          "description": "AccessKey",
          "type": "string"
        },
        "bucket": {
          "description": "Bucket",
          "type": "string"
        },
        "endpoint": {
          "description": "Endpoint",
          "type": "string"
        },
        "key": {
          "description": "Key",
          "type": "string"
        },
        "secretKey": {
          "description": "SecretKey",
          "type": "string"
        }
      }
    },
...
    "TaskStatusResponse": {
      "type": "object",
      "properties": {
        "created": {
          "description": "创建时间",
          "type": "string"
        },
        "finished": {
          "description": "完成时间",
          "type": "string"
        },
        "message": {
          "description": "消息",
          "type": "string"
        },
        "outputUrl": {
          "description": "输出URL",
          "type": "string"
        },
        "priority": {
          "description": "优先级",
          "type": "integer"
        },
        "progress": {
          "description": "进度",
          "type": "number"
        },
        "started": {
          "description": "开始时间",
          "type": "string"
        },
        "status": {
          "description": "状态",
          "type": "string"
        },
        "taskId": {
          "description": "任务ID",
          "type": "string"
        }
      }
    },
    "VideoEditRequest": {
      "type": "object",
      "properties": {
        "ossOutput": {
          "description": "OSS输出配置",
          "$ref": "#/definitions/OSSOutput"
        },
        "outputPath": {
          "description": "输出路径",
          "type": "string"
        },
        "priority": {
          "description": "任务优先级",
          "type": "integer"
        },
        "spec": {
          "description": "任务规格",
          "type": "object"
        }
      }
    },
    "VideoEditResponse": {
      "type": "object",
      "properties": {
        "message": {
          "description": "消息",
          "type": "string"
        },
        "outputUrl": {
          "description": "输出URL",
          "type": "string"
        },
        "status": {
          "description": "状态",
          "type": "string"
        },
        "taskId": {
          "description": "任务ID",
          "type": "string"
        }
      }
    },
    "VideoURLRequest": {
      "type": "object",
      "properties": {
        "url": {
          "description": "视频URL",
          "type": "string"
        }
      }
    },
    "VideoURLResponse": {
      "type": "object",
      "properties": {
        "error": {
          "description": "错误信息",
          "type": "string"
        },
        "message": {
          "description": "消息",
          "type": "string"
        },
        "status": {
          "description": "状态",
          "type": "string"
        },
        "taskId": {
          "description": "任务ID",
          "type": "string"
        },
        "tsFilePath": {
          "description": "TS文件路径",
          "type": "string"
        }
      }
    }
  }
}